pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestra"
version = "0.1.0"
description = "Local file-based multi-agent coding orchestrator."
requires-python = ">=3.11"
dependencies = [
    "jsonschema>=4.21",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
orch = "orch.cli:main"

[tool.setuptools.packages.find]
include = ["orch*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
